+++ lizard_map/context_processors/processor.py
@@ -2,8 +2,11 @@ import logging
 
 from lizard_map.animation import AnimationSettings
 from lizard_map.coordinates import MapSettings
+from lizard_map.daterange import DateRangeForm
+from lizard_map.daterange import current_start_end_dates
 from lizard_map.utility import analyze_http_user_agent
 from lizard_map.views import MAP_LOCATION
+from lizard_map.workspace import WorkspaceManager
 
 logger = logging.getLogger(__name__)
 
@@ -22,16 +25,11 @@ def detect_browser(request):
     return {'detected_browser': result}
 
 
-def processor(request):
-    """
-    A context processor to add the lizard_map variables to the current
-    context.
-    """
+def map_variables(request):
+    # Map variables.
     session = request.session
     add_to_context = MapSettings().map_settings
-    add_to_context['animation_slider'] = None  # default
 
-    # Map variables.
     # By default it takes coordinates from the django settings. If a
     # user has custom coordinates in his session, it will take those
     # coordinates and zoomlevel.
@@ -54,10 +52,40 @@ def processor(request):
             logger.error(
                 'Error fetching map coordinates from session: %s'
                 % map_location)
+    return add_to_context
+
+
+def workspace_variables(request):
+    add_to_context = {}
+
+    workspace_manager = WorkspaceManager(request)
+    add_to_context['workspaces'] = workspace_manager.load_or_create()
+
+    add_to_context['date_range_form'] = DateRangeForm(
+        current_start_end_dates(request, for_form=True))
 
     # Add animation slider? Default: no.
+    add_to_context['animation_slider'] = None  # default
     #animation_slider = AnimationSettings(request).info()
     #add_to_context['animation_slider'] = animation_slider
+
+    return add_to_context
+
+
+def processor(request):
+    """
+    A context processor to add the lizard_map variables to the current
+    context.
+    """
+    session = request.session
+
+    add_to_context = {}
+
+    # Add map variables.
+    add_to_context.update(map_variables(request))
+
+    # Add workspaces.
+    add_to_context.update(workspace_variables(request))
 
     # Add detected browser
     add_to_context.update(detect_browser(request))